+++ lib/src/controls/l10n/l10n.dart
@@ -1,0 +1,1 @@
+export 'flutter_deck_localization_notifier.dart';